Add files via upload
Q1. What is an ensemble technique in machine learning?

Q2. Why are ensemble techniques used in machine learning?

Q3. What is bagging?

Q4. What is boosting?

Q5. What are the benefits of using ensemble techniques?

Q6. Are ensemble techniques always better than individual models?

Q7. How is the confidence interval calculated using bootstrap?

Q8. How does bootstrap work and What are the steps involved in bootstrap?

Q9. A researcher wants to estimate the mean height of a population of trees. They measure the height of a
sample of 50 trees and obtain a mean height of 15 meters and a standard deviation of 2 meters. Use
bootstrap to estimate the 95% confidence interval for the population mean height.