Add files via upload
Q1. How does bagging reduce overfitting in decision trees?

Q2. What are the advantages and disadvantages of using different types of base learners in bagging?

Q3. How does the choice of base learner affect the bias-variance tradeoff in bagging?

Q4. Can bagging be used for both classification and regression tasks? How does it differ in each case?

Q5. What is the role of ensemble size in bagging? How many models should be included in the ensemble?

Q6. Can you provide an example of a real-world application of bagging in machine learning?