Add files via upload
Q1. What is the KNN algorithm?

Q2. How do you choose the value of K in KNN?

Q3. What is the difference between KNN classifier and KNN regressor?

Q4. How do you measure the performance of KNN?

Q5. What is the curse of dimensionality in KNN?

Q6. How do you handle missing values in KNN?

Q7. Compare and contrast the performance of the KNN classifier and regressor. Which one is better for
which type of problem?

Q8. What are the strengths and weaknesses of the KNN algorithm for classification and regression tasks,
and how can these be addressed?

Q9. What is the difference between Euclidean distance and Manhattan distance in KNN?

Q10. What is the role of feature scaling in KNN?