Add files via upload
Q1. What is the main difference between the Euclidean distance metric and the Manhattan distance
metric in KNN? How might this difference affect the performance of a KNN classifier or regressor?

Q2. How do you choose the optimal value of k for a KNN classifier or regressor? What techniques can be
used to determine the optimal k value?

Q3. How does the choice of distance metric affect the performance of a KNN classifier or regressor? In
what situations might you choose one distance metric over the other?

Q4. What are some common hyperparameters in KNN classifiers and regressors, and how do they affect
the performance of the model? How might you go about tuning these hyperparameters to improve
model performance?

Q5. How does the size of the training set affect the performance of a KNN classifier or regressor? What
techniques can be used to optimize the size of the training set?

Q6. What are some potential drawbacks of using KNN as a classifier or regressor? How might you
overcome these drawbacks to improve the performance of the model?