Add files via upload
Q1. What is a projection and how is it used in PCA?

Q2. How does the optimization problem in PCA work, and what is it trying to achieve?

Q3. What is the relationship between covariance matrices and PCA?

Q4. How does the choice of number of principal components impact the performance of PCA?

Q5. How can PCA be used in feature selection, and what are the benefits of using it for this purpose?

Q6. What are some common applications of PCA in data science and machine learning?

Q7.What is the relationship between spread and variance in PCA?

Q8. How does PCA use the spread and variance of the data to identify principal components?

Q9. How does PCA handle data with high variance in some dimensions but low variance in others?